finalize sync-click-workspaces cron
