+++ apps/web/app/(ee)/api/cron/webhooks/sync-click-workspaces/route.ts
@@ -1,9 +1,9 @@
 import { withCron } from "@/lib/cron/with-cron";
-import {
-  promoteLinkWebhooksForClick,
-  syncClickWebhookWorkspaceSet,
-} from "@/lib/webhook/click-webhook-workspaces";
 import { logAndRespond } from "../../utils";
+import {
+  cleanupRedundantLinkWebhookEntries,
+  syncClickWebhookWorkspaceSet,
+} from "./utils";
 
 export const dynamic = "force-dynamic";
 
@@ -11,11 +11,10 @@ export const dynamic = "force-dynamic";
 // Rebuild the Redis set of workspaces with active link.clicked webhooks.
 // Runs every minute (* * * * *)
 export const GET = withCron(async () => {
-  const promoted = await promoteLinkWebhooksForClick();
-
   const synced = await syncClickWebhookWorkspaceSet();
+  const deleted = await cleanupRedundantLinkWebhookEntries();
 
   return logAndRespond(
-    `Synced ${synced} workspace(s) with link.clicked webhooks. Promoted ${promoted} webhooks for click.`,
+    `Synced ${synced} workspace(s) with link.clicked webhooks. Deleted ${deleted} redundant LinkWebhook entries.`,
   );
 });
